Move storage option parsing to utilities
